Rename the helm mixin to helm2
This removes "helm" from the results of `porter mixins search` and
ensures that it shows up as helm2 instead. If someone was installing the
mixin with `porter mixins install helm` that will continue to work
because we have old entries in our atom.xml but they will get the
last version of the helm mixin before we renamed it.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>